+++ src/components/home/HomeHeader/HomeHeader.tsx
@@ -3,12 +3,12 @@ import TitleDate from './TitleDate/TitleDate';
 
 import React from 'react';
 
-interface HomeHeaderContainerProps {
+interface HomeHeaderProps {
   /**시트 열기 위한 함수 */
   handleSetOpen: (value: boolean) => void;
 }
 
-const HomeHeaderContainer: React.FC<HomeHeaderContainerProps> = ({ handleSetOpen }) => {
+const HomeHeader: React.FC<HomeHeaderProps> = ({ handleSetOpen }) => {
   return (
     <div className='flex items-center justify-between bg-white02 p-5'>
       <TitleDate dateText='2024년 11월 둘째 주' />
@@ -17,4 +17,4 @@ const HomeHeaderContainer: React.FC<HomeHeaderContainerProps> = ({ handleSetOpen
   );
 };
 
-export default HomeHeaderContainer;
+export default HomeHeader;
