+++ src/components/home/HouseworkList/HouseworkList.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import ListContainer, { ListContainerProps } from './ListContainer';
+import ListContainer, { ListContainerProps } from './HouseworkList';
 
 const meta = {
   title: 'components/ListContainer',
